WEB-156: [bugfix] search_options/select collision
