+++ src/search/search.js
@@ -1,19 +1,8 @@
-import {LitElement, html, css} from 'lit-element/lit-element'; //'https://unpkg.com/lit-element@2.2.0/lit-element.js?module';
+import {LitElement, html, css} from 'lit-element/lit-element';
+import {search_options} from './search_options';
+
 const ENTER_KEY_VALUE = 13;
 const default_to_just_primo = true;
-
-/** data on the overall search sources we have available to search on */
-export const search_options = [
-  {"value":"help",      "name":"Ask a Librarian",          "placeholder": "Search for help topics",      "domain":"http://askalibrarian.bu.edu/search/?t=0&q="},
-  {"value":"primo",     "name":"BU Libraries Search",      "placeholder": "Search library resources",     "domain":"https://buprimo.hosted.exlibrisgroup.com/primo-explore/search?vid=BU&institution=BOSU&search_scope=default_scope&highlight=true&lang=en_US&query=any,contains,"},
-  {"value":"wp",        "name":"BU Libraries Site",        "placeholder": "Search library info/services", "domain":"https://search.bu.edu/?site=www.bu.edu%2Flibrary&q="},
-  {"value":"industries","name":"Industry Survey Locator",  "placeholder": "Search for industry surveys",  "domain":"https://buprimo.hosted.exlibrisgroup.com/primo-explore/search?vid=ISL&institution=BOSU&search_scope=default_scope&highlight=true&lang=en_US&query=any,contains,"},
-  {"value":"guides",    "name":"Library Guides",           "placeholder": "Search Research Guides",       "domain":"http://library.bu.edu/srch.php?q="},
-  {"value":"openbu",    "name":"Open BU",                  "placeholder": "Search BU Digital Collections","domain":"https://open.bu.edu/discover?query="},
-  {"value":"worldcat",  "name":"OCLC WorldCat",            "placeholder": "BU Libraries Search",          "domain":"https://bu.on.worldcat.org/search?queryString="}
-  // {"value":"directory", "name":"Staff Directory",          "placeholder": "Search for people at BU",      "domain":"https://www.bu.edu/phpbin/directory/?q="},
-  // {"value":"hgar",      "name":"Archival Research Center", "placeholder": "Search the BU Archive",        "domain":"http://archives.bu.edu/search/?search=SEARCH&query="},
-];
 
 const search_option_codes = search_options.map( x => x.value );
 
@@ -67,6 +56,7 @@ export default class BULibSearch extends LitElement {
           background-color: #212121;
           background: linear-gradient(#212121,#333333,#444444);
           border-radius: 4px;
+          font-family: 'Benton', sans-serif !important;
         }
         #bulib-search { 
           display: flex;
